Add tabstop option to Component.Linkifier
This CL adds a tabstop option to the various linkify() functions under
Component.Linkifier. Allowing Linkifier users to choose whether they
want a tabbable link or not is neccessary to satisfy accessibility
requirements without adding extra tabstops to links that are made
accessible via a context menu or some other alternative. As some of the
argument lists were getting long, I also moved the optional arguments
into an object.

These changes will break a web test, so [1] must be merged first. The
fixes are in [2]

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1894280
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1721843

Bug: 963183
Change-Id: Iaa3047f5fffa609c9e5c0b6cb49ccbeee5737de1