ListControl ARIA improvements and opt-out
This CL marks UI.ListControls as listboxes and their list items as
options, in accordance with the WAI-ARIA listbox pattern at [1]. In
order to allow ListControl users to opt out of this ARIA management,
it also adds a new updateSelectedItemARIA() method to UI.ListDelegate
which prevents ListControl from setting aria-selected on selected items.
Additionally, if the element returned by createElementForItem already
has a role, UI.ListControl will not change it. This allows users to mark
the items as list items instead of options, which is necessary for the
breakpoints panes.

Other changes:
- calls preventDefault in the keydown handler to prevent arrowing
through the list from unnecessarily scrolling
- throws an error instead of a string (requested by Joel in previous CL)

This CL breaks a test, so [2] should be merged first to disable it. The
fixes for the test are at [3].

Originally reviewed at https://chromium-review.googlesource.com/c/chromium/src/+/1690089

[1] https://www.w3.org/TR/wai-aria-practices/#Listbox
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1881715
[3] https://chromium-review.googlesource.com/c/chromium/src/+/1690089

Bug: 963183
Change-Id: I2a81744371497d0e592556f65f0c8177826d7c19