Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

I think the byte values might be a bit redundant, but it is useful in the sense that it provides more context and is a more complete sentence. I think this is a decision between developers hovering on "XX KB / XXX KB transferred" to see the explanation of this whole phrase, vs. the explanation of just the word "transferred". Also, it seems like many popup texts in Network panel show the original hovered texts, so for now I would stay with the tradition and be a bit redundant.

Patch-set: 3