Fix Settings screen blocking interaction with Device Toolbar
This tiny PR addresses a bug with the Device Toolbar when the Settings
dialog is open.

Originally, the Settings dialog had behavior where clicking outside of
the Settings dialog caused it to close. This originally caused the NVDA
screen reader to accidentally dismiss the dialog when buttons were
pressed. That behavior was disabled in this CL.

https://chromium-review.googlesource.com/#/q/Ie1773ed8987dd6d42f8c9cb897fea3de5a091bcb

However, it introduced a new issue with the Device Toolbar interaction.
The Device Toolbar is a region of the DevTools that, when docked, is
part of the same document as the DevTools. Before the change, users
that had the Settings dialog open could click on the Device Toolbar
once to dismiss the Settings dialog (as it is outside the dialog), and
then click a *second time* to interact with the Device Toolbar.

With the change in the above CL, the *first click* no longer dismisses
the Settings dialog, preventing user interactions with the Device
Toolbar on the *second click*.

The change in this CL enable clicks outside the Settings dialog to
"pierce" the GlassPane that overlays the document, so that users may
interact with the Device Toolbar on the *first click* while *also*
not dismissing the Settings. This improves the experience all around:

 - Screen reader users do not accidentally dismiss Settings
 - Users who want to add devices can have the Settings pane and Device
   Toolbar open simultaneously
 - Users may interact with the Device Toolbar on the *first click*
   rather than being forced to dismiss the Settings screen first

Before: (see videos in bug below)

https://bugs.chromium.org/p/chromium/issues/detail?id=1023755

After: https://i.imgur.com/47WVpgV.gif

Bug: 1023755
Change-Id: I1828235fb7db78352a2926b31c92f218be154eb9