Fix autocomplete TypeError: protoNames is not iterable
Background:
protoNames is needed to display JavaScript object properties in the
autocomplete flyout. For example when editing code snippets in the
console, watches, and source pane.

Problem:
When callFunctionJSON fails due to a Protocol Error protoName is null [1]
which will cause the runtime TypeError: protoNames is not iterable [2].

Solution:
Check that protoNames is truthy before iterating over it; return early
with an empty array if protoNames is falsy.

[1]: https://cs.chromium.org/chromium/src/third_party/devtools-frontend/src/front_end/sdk/RemoteObject.js?l=658&rcl=6ced335a
[2]: https://cs.chromium.org/chromium/src/third_party/devtools-frontend/src/front_end/object_ui/JavaScriptAutocomplete.js?l=172&rcl=6ced335

Change-Id: I1431d3de6333bfd331e8ec1a1446d8051aa24157