Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I have a weird interaction:
> > 
> > https://imgur.com/a/LQjdR4F
> > 
> > It looks to me like "Frames" is selected, but when I arrow-right, "Frames" scrolls out of view and "Network" is still present. Can you help me understand what's going on?
> 
> Network and Frames are part of 2 different flamecharts. I drew an outline around the selected flamechart, but it may be hard to see. The flamechart with Network row may or may not exist, which is why I think it is a separate entity.

OK, we had a conversation offline and I think we've stumbled onto a few things that are suspect.

1. When I expand "Network", the left-side selector tick obviously doesn't cover the full height of the row. I could see this because the background color of the row bled past the bottom of the tick, as illustrated in the following exaggerated ASCII art:

____
|
----
(background color of Frames row)

This looks like just an off-by-one error and is probably straightforward to address.

2. Possibly related to the former and possibly not: When expanding Frames, Main, or another item that is larger than the vertical space afforded to it, it scrolls to the bottom of that row. It's weird here because when I do so, the selection tick for "Network" and the selection tick for the row that is selected bleed together, like this:

________
|   Network
|      (row for, say, Main, but its label isn't visible
|
|
|---------

This one is VERY confusing -- it made me think multiple times that I'd expanded the wrong row. I believe the correct behavior would be to not scroll to bottom in cases where the expanded row would be bigger than the available viewport. For example:

-----------
|   Network
|----------
|   Main
|
|

Additionally, in these kinds of circumstances, you might consider scrolling the entirety of the viewport, so that Network is NOT visible.

Patch-set: 1
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>