Add TypeScript Ninja integration
This CL adds integration for TypeScript in the Ninja build system. It includes the ts_library.py script that calls the typescript compiler on the source files. The Ninja template in typescript.gni is a similar definition to the Bazel definition [1]. It does not yet have dependency resolution, which will be implemented in a follow-up CL. The tsconfig will also change later, but that is still up for discussion.

[1]: https://bazelbuild.github.io/rules_nodejs/TypeScript.html#compiling-typescript-ts_library

Allow changes in the .gitignore for new third_party inclusions

Otherwise the upload fails as it would include a change in both a `third_party` directory as well as the `.gitignore`.

Bug:1011811
Change-Id: I11a85318cfcc791cb277388aed6028b34f749a14