Teach Closure about TypeScript files
With the introduction of the `typescript` namespace in `module.json`, we can teach Closure about files that are generated by TypeScript.

As such, we will have 3 different type-checkable files:
1. Files that are solely type-checked by Closure.
2. Files that are solely type-checked by TypeScript.
3. Files that are type-checked by both Closure and TypeScript.

Files of type 1 are only allowed to import files from type 1 and 3.
Files of type 2 are only allowed to import files from type 2.
Files of type 3 are only allowed to import files from type 2.

Files of type 3 must include definitions for custom elements and declare their definition in the file itself.
Files of type 3 are allowed to import files from type 2 and are the files that should be included in the `typescript` array in the `module.json`.

This approach will minimize the compatibility of the two type-systems.
By importing and relying on the side-effect of the `customElements.define` call, the `document.createElement` calls in files of type 2 will be "ignored" by Closure.
(Ignored in the sense that Closure does not type-check the subclasses of HTMLElements created by `createElement`, as it infers the result of the `createElement` call to be just `HTMLElement` instead of a specific subclass.)
By also running TypeScript over these files, we can actually ensure type-safety of any interaction with the custom element.
TypeScript *does* type-check any interaction with a custom element, for example with property assignment or method invocations.

Change-Id: I84e00cf4455444b5bf1b536df6d9641dcbe9c8d3