Update patch set 1
Patch Set 1:

I tried the change, in general it LGTM.

I'm not crazy about all of the individual assignments. I understand that's kind of exchanging the currencies you have on "either side" if that makes sense, but it's looks headache-inducing.

Can you take a look and see if there's a better way to write this so that we don't have a bunch of ternaries?

Thanks!

Patch-set: 1
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>