Use rollup to bundle module entrypoints
There was a performance regression in terms of number of files loaded,
which was running into the Chromium network loading limit of 6
concurrent files. By bundling the respective entrypoints of each
DevTools module, we reduce the total number of files loaded and
therefore should once again improve the loading of performance of
DevTools as a whole.

It modifies the build_release_application to bundle all files in a
respective module in its entrypoint, while leaving relative imports to
sibling module entrypoints alone.

For example, if `common/common.js` imports `../platform/platform.js`,
then `platform.js` is NOT bundled in `common.js`. The reason is that we
need to maintain compatibility with the old module system, where we
don't have the full import-dependency graph yet.

Bug:1006759
Change-Id: Idc01b73008de41e28c83e4df789f57c0b954e71a