[DevTools]: Always show contrast checker UI.
Contrast Details is currently shown based on whether there is font & background
color info available for the inspected element. This patch defaults the font
size to 'normal' and allows the developer to use the pre-existing color picker
UI to set the background color of their choosing.

This is patch 1 of 2. The other patch will be applied in
devtools/devtools-frontend once this lands.

Bug: 956711
Change-Id: I6ffd485aa02ea547689512c6d182ed573d626250