Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can we use a Gerrit mirror for this, rather than checking in all source files?

I intend to make modifications to this to get it to return a mapping between the disassembly and byte offsets. I don't know if this would be easy to upstream.

> 
> Also, when you include this package, what is the impact on the bundle size?

This gets compiled to a 1.4 MB JavaScript file.

> 
> Lastly, has this been security reviewed?

No. This gets compiled to a webassembly module which we access through a JS API. Does this need a security review?

Patch-set: 1
Reviewer: Gerrit User 1369666 <1369666@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0