Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Can we use a Gerrit mirror for this, rather than checking in all source files?
> 
> I intend to make modifications to this to get it to return a mapping between the disassembly and byte offsets. I don't know if this would be easy to upstream.

Per the Google-internal third_party policy, it is actively discouraged to make local modifications to third_party packages. It would be better to upstream the changes. If upstream does not accept the patches, we should investigate other options. For example, we could augment the API with some DevTools-specific implementation.

> 
> > 
> > Also, when you include this package, what is the impact on the bundle size?
> 
> This gets compiled to a 1.4 MB JavaScript file.

That is a lot of code. We can't ship that by default in DevTools. What is the technical reason for needing this package? Are there alternatives that are more lightweight? I would like to not introduce more remote modules, so any solution that does not have to be made remote would be best.

> > 
> > Lastly, has this been security reviewed?
> 
> No. This gets compiled to a webassembly module which we access through a JS API. Does this need a security review?

Yes, any third_party package needs review per the Chromium third_party policy. It will likely be easy to obtain approval given what you said, but we need approval nonetheless.

Patch-set: 1
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>