Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Fine by me if that's what we really want (we seem to propagate runtime errors to `console.error` quite frequently in the codebase in general).

I feel like the devtools team doesn't have any influence on whether or not people have propper sourcemap support on their website. Therefore cluttering the console of the devtools inspecting devtools seems not too helpful. But I don't know if there are any usecases where it would be useful though.

The cl still prints a warning to the console that is inspecting the website...

Patch-set: 1