Update patch set 1
Patch Set 1:

> > 
> > > 
> > > Also, when you include this package, what is the impact on the bundle size?
> > 
> > This gets compiled to a 1.4 MB JavaScript file.
> 
> That is a lot of code. We can't ship that by default in DevTools. What is the technical reason for needing this package? Are there alternatives that are more lightweight? I would like to not introduce more remote modules, so any solution that does not have to be made remote would be best.

The technical reason is that we are making disassembling wasm for debugging the responsibility of the frontend rather than V8 so that we can get rid of behavior where V8 is sometimes doing some source mapping behind the scenes.

I haven't found any existing JavaScript Wasm to Wat conversion libraries. So the alternative is to write one. The C++ code this is replacing (mostly wasm-text.cc and wasm-serialization.cc) is maybe 50K, so a pure JavaScript solution shouldn't have to be large, though it will take longer to write and it may be slower.

Patch-set: 1
Reviewer: Gerrit User 1369666 <1369666@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0