Sensor: Add error message for timezone validity
This change adds error message for timezone validity and alert user both visually and with screen reader message when invalid inputs are entered.
This error validation in Sensor tool for timezone in Geolocation section is same as timezone validation in Geolocation Settings which is linked by 'Manage' button in Sensor tool.

Gif showing the change: https://imgur.com/a/qBo6Iah
Bug: 963183
Change-Id: I2514f5ce6e0cb23fa3d7992f449659e68d184c21