Prevent Settings dialog from dismissing when using other tools
This CL changes the behavior of the Settings dialog so users can
simultaneously interact with Settings and the Device Toolbar (or the
Screencast app).

The CL disables the default click-outside-to-dismiss behavior of the
Settings dialog and instead permits clicks to penetrate the outer Glass
Pane element that wraps the dialog. Furthermore, Dialog now allows for
the preservation of tab index values outside the Inspector.

This CL is a work in progress and is not ready to be merged.

Bug: 963183
Change-Id: I6e650c0a3fbba071a67cde327acda6c6f2a574a0