[DevTools]: Always show contrast checker UI.
Contrast Details is currently shown based on whether there is font & background
color info available for the inspected element. This patch defaults the font
size to 'normal' and allows the developer to use the pre-existing color picker
UI to set the background color of their choosing.

This is patch 2 of 2 to be landed after https://crrev.com/c/chromium/src/+/1932786

Bug: 956711
Change-Id: Ibee948f8124bc7ae786e596e1d5de990c180697c