Fix `npm test` when run from the devtools/devtools-frontend repo
The script previously assumed it would always be run from within the
chromium/src/third_party/devtools-frontend/src folder. It would fail
when run from the standalone repository at devtools/devtools-frontend.

This patch makes it work even in that case, as long as the `devtools`
directory in `devtools/devtools-frontend` lives on the same level as
the `chromium` repository.

Change-Id: I3cf81c8144fd3f03389a4ec395b8ec3acdbb8e55