Datagrid: Make the header context menu keyboard accessible
This change adds a sub menu item to the datagrid context menu that
allows the Header context menu to be accessible from the
contextMenuKey press in the body of the datagrid. This is similar
to the "Sort By" menu option added previously for the same
purpose.

Currently, the header options are used in the network tool and the
protocol monitor.

Before: https://imgur.com/oGh9fwr

After: https://imgur.com/G37Cts3

Bug: 963183
Change-Id: I4d02bed0dc9924be7fd415cad4053711a3aeb044