Update Googlebot User-Agent
webmasters.googleblog.com/2019/10/updating-user-agent-of-googlebot.html
describes that Google is updating their Googlebot User-Agent to report
the Chrome/%s version in (which will roughly update at the same time as
Chrome stable).

So devtools' Googlebot UA shouldn't used a fixed version either. Replace
it with %s so it matches the version number Chrome was built with.

This won't always exactly match the UA used by Google, but is probably
the closest we can reasonably get.

While I'm here I also add a third "Googlebot Desktop" User Agent
to mirror the smartphone one. (See "Googlebot user agents today" in the
above link)

Note that I plan to submit this change without coordinating the timing
with the internal UA change rollout. This means there will be a period
of time where the UA in devtools doesn't match the UA used by google.

Change-Id: I68a0d5549fab70d0ce8e37e9f518b5a0aead68eb