DevTools: Add profileType context in the aria-description of Memory tool sidebar tree elements
When the screenreader read out the tree element labels for Memory tool sidebar, it only read out content of the tree element and lacked the context of the profile type that the tree element was parented under. Added that as aria-description, now it reads out Profile-Type alongside the contents of the tree element.

Pictures: https://imgur.com/a/RztWrpU

Bug: 963183
Change-Id: I557cd725a68eb993969aade6b6d8a0934873ac02