Update patch set 9
Patch Set 9:

This is ready to land but is failing b/c of layout tests in Chromium asserting on lighthouse version # in devtools frontend. I want to move those layout tests to frontend layout tests but the infra is not there yet.

I don't think there's a rush to land this, so I'll just wait for the infra to be ready for layout tests.

Patch-set: 9
Reviewer: Gerrit User 1378407 <1378407@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0