Do not reformat `1 .toString()` to `1.toString()`
Previously devtools JS formatter assumed that it would never need a
space between the object and the dot in a MemberExpression. This
was incorrect for integer literals like in `1 .toString()`, as
after formatting it as `1.toString()` the `1.` will be parsed as
part of the integer instead of as a member expression.

This means we either need to put a space between integer and dot,
or add an extra dot.

This change leaves the space by checking if the LHS of a
MemberExpression is an integer number literal. Note this will
also unnecessarily add a space in cases like "0x1.toString()".

Note: This change will be tested in a follow-up change by updating
the text-editor-formatter.js layout test. Not done here since it's
in a different repository.

Bug: 1031276
Change-Id: I7fa271caea00192e199c3724386babaa8d2a99ae