Improve styling for maskable icon functionality
Previously, the checker background would get the height of the field
description if that was larger than the image itself.

With this patch, the checker background only applies to the correct
image sizes, even for small icons.

The patch also improves the spacing between multiple icons in case
they are shown side-by-side (i.e. the manifest defines multiple icons
with the same size).

Before/after screenshots:
https://bugs.chromium.org/p/chromium/issues/detail?id=1027658#c9

Bug: chromium:1027658
Change-Id: Ic6b1352a984031157004e4a99a33a54f30d73d8b