Update patch set 4
Patch Set 4:

(2 comments)

> Patch Set 4:
> 
> (3 comments)
> 
> The wasmparser is a front_end third_party dependency and must therefore be added to `front_end/third_party`. Could you follow the guidelines (e.g. security approval et al) and add it in a separate CL?

Will do. Do we want a full copy of the original typescript just for security review purposes, even if we can't link to it as is? I'm following the acorn approach of checking in the built files where they are being used, though I see that library was checked into third_party as well.

Patch-set: 4