Update patch set 6
Patch Set 6:

(2 comments)

> Patch Set 6:
> 
> (2 comments)
> 
> You're seeing the results of me having my weird habit of reviewing changes from the bottom upwards. 😊
> 
> Am I understanding this correctly that you're not meaning to use TypeScript to compile? I don't think that this will really give us a lot of value. Using TypeScript compiler to typecheck Closure annotations is likely to be a step backwards for type verification.

No, we have the intention to use TS as the compiler. However, we can't do that in 1 go, as the changes would be too large and unreviewable. This is a small step to a TypeScript world, but not the end result we want to eventually have.

> As you noted, TypeScript doesn't support @interface. It has a number of other ways to express these kinds of structural types (which is what TypeScript is oriented around anyway) in interfaces, type aliases, unions, etc. However, these type expressions are a semantic part of the TypeScript language, and I'm fairly certain they need to be contained in .ts files (because the types are erased as part of the compilation process).

We can't go full TypeScript straight away, as it would not integrate with the old code. E.g. we need to type-check that the old code uses the new code correctly. I called these "bridges" in my comment below. For that, TypeScript will need to check a minimal amount of JavaScript which uses TypeScript.

I realize that this CL might come out of the blue, but it is hard for me to test what the capabilities of TypeScript are and how it could integrate without submitting CLs and iterating on it. The design doc for TypeScript will be the output of these experiments about what works/doesn't work.

I am not sure what the best way forward would be, as I will need to land CLs and prototype, but I also want to make clear that nothing is set in stone. I tagged this CL as [EXPERIMENTAL] to hopefully make that clearer.

Patch-set: 6
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>