Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (1 comment)
> 
> I totally get that we can't directly jump to TypeScript (and the experimental moniker) -- just offering to help. What I see is that we have to solve two problems:

Yes, this distinction is helpful! This CL intended to only solve 1 small step towards a world where we can solve problem 2. It also provides a small step towards problem 1, but that would require additional build system plumbing I haven't had the time for.
> 
>  1. Where do "built" JavaScript files get emitted so that they can be picked up and packaged?
>  2. How do we integrate new TypeScript with existing JavaScript code that is using JSDoc?
> 
> For the latter, which is what I think you're trying to do right now: it might be worth looking into what we can do with the exiting code. We can check in generated .d.ts files so that they're just available, and can even consider including a presubmit step that might allow us to update the declarations ahead of time.

Most of the time, the Closure types are wrong. This is what Paul have been hitting constantly throughout the ES modules migration. TypeScript finds out that the types are wrong, which means that it would force us to fix old types if we want to teach the new TypeScript world about the old world. I expect that the additional work of trying to obtain clean `.d.ts` files would be substantial. Moreover, it would be a manual process of generating these `.d.ts` files, which results in my point below.

> I'm confident that we can do (2) pretty readily as incremental updates to the existing code (it's been something I've done several times). I (personally) am far more "scared" of (1) -- I'm still learning my way around ninja, and my experience has shown that changes like that tend to impact a lot more other scripts and the like.

The BUILD.gn would generate files in the gen folder relative to the `.ts` definitions. I have a plan there, but will need to make some changes. I am quite confident the plan will work, but it will just require some plumbing in the build system. Rest assured, this problem will be tackled 😊
 
> Coming back to (2) -- I think there are going to be two directions to solve this:
>  -- How do we provide type information from old JS to new TS? I think a stop-gap for that will be in leveraging the ambient .d.ts files, and we can manually add annotations as we start writing new code

This was the original solution I had in mind (https://chromium-review.googlesource.com/c/chromium/src/+/1728584), but we quickly realized this does not scale. At that point, we are lying to the TypeScript compiler. If someone makes a change that changes the types, we would have to keep them in sync. As far as I know, there is no automated check that can make that happen. I don't think that solution scales for the codebase of DevTools and the expect time we will have to live with both type checkers.

>  -- How do we provide type annotations from TS to old code? This is a bit more involved, but in general, that will be the direction for renaming .js files to .ts.

I am not sure if I follow. Many Typescript language features can not be mapped to JSDoc (e.g. interfaces, generics, etc...). Until we remove our dependency of Closure, we will always have to artificially limit ourselves using TypeScript features?

> I think it's likely that Closure and TypeScript will live alongside each other for a while while we go through this process. Happy to help however I can.

Patch-set: 6
Reviewer: Gerrit User 1365394 <1365394@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>