Update patch set 3
Patch Set 3:

(1 comment)

Additionally: When actually using NVDA, cannot find a mechanism to read some queries (I was able to get it to read an 'insert' statement but not a 'select' statement), and I could not get back to a point using only the arrows to enter a new query. Tabbing out and back in might have worked, but the arrow-down behavior should allow me to get there and it could not.

Patch-set: 3
Reviewer: Gerrit User 1329477 <1329477@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>