Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)

I would say for most things, we have just added the entries to the context menu without filtering for context menu key. Examples include: links in datagrid and console, the sort by option for datagrids. This case is unique as the fuctionality is available in a different context menu for mouse user already (right clicking on the header).

I do think it would improve discoverability and consistency to not filter for this case though, so it's always available. (I don't know if many people know about the existing menu). I'll make the update

Patch-set: 5